Struct generic_vec::raw::Heap

#[repr(C)]
pub struct Heap<T, A: ?Sized + Allocator = Global> { /* fields omitted */ }

This is supported on crate feature alloc only.

A heap storage that can reallocate if necessary,

Usable with the alloc feature

The allocator type paramter is only available on nightly

Implementations

impl<T> Heap<T>

pub const fn new() -> Self

Create a new zero-capacity heap vector

pub const unsafe fn from_raw_parts(ptr: NonNull<T>, capacity: usize) -> Self

Create a new Heap<T>storage from the given pointer and capacity

Safety

If the capacity is non-zero

You must have allocated the pointer from the Global allocator

The pointer must be valid to read-write for the range ptr..ptr.add(capacity)

pub const fn into_raw_parts(self) -> (NonNull<T>, usize)

Convert a Heap storage into a pointer and capacity, without deallocating the storage

impl<T, A: Allocator> Heap<T, A>

pub const fn with_alloc(allocator: A) -> Self

This is supported on crate feature nightly only.

Create a new zero-capacity heap vector with the given allocator

pub const unsafe fn from_raw_parts_in(
ptr: NonNull<T>,
capacity: usize,
allocator: A
) -> Self

This is supported on crate feature nightly only.

Create a new Heap<T>storage from the given pointer and capacity

Safety

If the capacity is non-zero

You must have allocated the pointer from the given allocator

The pointer must be valid to read-write for the range ptr..ptr.add(capacity)

pub fn into_raw_parts_with_alloc(self) -> (NonNull<T>, usize, A)

This is supported on crate feature nightly only.

Convert a Heap storage into a pointer and capacity, without deallocating the storage