Struct generic_vec::raw::UninitSlice

#[repr(transparent)]
pub struct UninitSlice<T>(_);

An uninitialized slice storage

Implementations

impl<T> UninitSlice<T>

pub const fn from_mut(buffer: &mut [MaybeUninit<T>]) -> &mut Self

Create a new UninitSlice storage

Note: this is only const with the nightly feature enabled

pub const unsafe fn to_mut(&mut self) -> &mut [MaybeUninit<T>]

Get the backing value of the this Uninit storage

Note: this is only const with the nightly feature enabled

Safety

You may not write uninitialized memory to this slice